# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LEIZ STORE API
  description: |
    REST API backend leiz-store (Next.js 14 App Router + Supabase PostgreSQL).
    Spesifikasi ini diturunkan dari `docs/api-specification.md` dan kode sumber
    di `src/app/api/**`.

    ## Konvensi
    - Semua endpoint (publik, admin, system) memakai envelope standar
      `{ success, data, meta }` untuk sukses dan `{ success, error: { code, message, details } }`
      untuk gagal. List endpoint menempatkan array di `data` dan paginasi di `meta`.
    - Semua error dikembalikan via `error.code` konsisten
      (`VALIDATION_ERROR`, `UNAUTHORIZED`, `FORBIDDEN`, `NOT_FOUND`, `CONFLICT`,
      `RATE_LIMITED`, `INTERNAL_ERROR`).

    ## Autentikasi
    - **Admin:** Supabase Auth (cookie `sb-*` atau `Authorization: Bearer <access_token>`)
      ATAU legacy JWT (cookie `admin_token` atau Bearer HS256, role `ADMIN`, 24 jam).
    - **System:** header `x-admin-secret` (`DISCORD_ADMIN_SECRET`) untuk
      `POST /api/admin/payment-action`; header `x-cron-secret` / `Authorization: Bearer`
      (`CRON_SECRET`) untuk cron.
    - **Discord:** verifikasi signature ed25519 (`x-signature-ed25519`,
      `x-signature-timestamp`) dengan `DISCORD_PUBLIC_KEY`.
  version: 1.0.0
  contact:
    name: LEIZ STORE Backend
  license:
    name: Proprietary
servers:
  - url: https://leiz-store.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
...
security: []
paths:
...
  /api/admin/products/{id}:
    put:
      tags: [Admin]
      summary: Update produk
      operationId: adminUpdateProduct
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpsertRequest" }
      responses:
        "200":
          description: Produk diperbarui (envelope, data = { product, message }).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          product: {}
                          message: { type: string }
              example:
                success: true
                data:
                  product:
                    id: "prod_1"
                    name: "500 UC"
                    slug: "500-uc"
                    price: 100000
                    is_active: true
                    category_id: "cat1"
                  message: "Product updated successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: Slug sudah dipakai produk lain.
    delete:
      tags: [Admin]
      summary: Nonaktifkan produk (soft delete)
      description: Menyetel `is_active = false` untuk menjaga riwayat order.
      operationId: adminDeleteProduct
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Produk dinonaktifkan.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message: { type: string, example: "Produk berhasil dinonaktifkan" }
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
...
  /api/admin/categories/{id}:
    put:
      tags: [Admin]
      summary: Update kategori
      operationId: adminUpdateCategory
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryUpsertRequest" }
      responses:
        "200":
          description: Kategori diperbarui.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          category: { $ref: "#/components/schemas/Category" }
        "404":
          $ref: "#/components/responses/NotFoundError"
    delete:
      tags: [Admin]
      summary: Hapus kategori
      description: Gagal (400) bila kategori masih memiliki produk.
      operationId: adminDeleteCategory
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Kategori dihapus.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message: { type: string, example: "Category deleted successfully" }
        "400":
          description: Kategori masih memiliki produk.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
              example:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: "Cannot delete category with 5 product(s). Reassign or delete products first."
...
  /api/admin/settings:
    get:
      tags: [Admin]
      summary: List pengaturan toko
      operationId: adminGetSettings
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Daftar setting (envelope, data = array).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Setting" }
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
    put:
      tags: [Admin]
      summary: Upsert pengaturan
      operationId: adminUpsertSetting
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertSettingRequest" }
            example: { key: "store_name", value: "LEIZ STORE", type: "text", group: "general" }
      responses:
        "200":
          description: Setting disimpan (upsert).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          setting: { $ref: "#/components/schemas/Setting" }
                          message: { type: string, example: "Setting updated successfully" }
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: |
        Cookie session admin. Menerima `admin_token` (legacy JWT) dan cookie
        Supabase Auth (`sb-*`).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token admin — access token Supabase Auth ATAU legacy JWT HS256
        (role `ADMIN`).
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
      description: Secret `DISCORD_ADMIN_SECRET` untuk aksi verifikasi pembayaran.
    cronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: "Secret `CRON_SECRET` untuk endpoint cron. Alternatifnya `Authorization: Bearer <secret>`."
    discordSignature:
      type: apiKey
      in: header
      name: x-signature-ed25519
      description: |
        Signature ed25519 Discord (dengan header `x-signature-timestamp`),
        diverifikasi terhadap `DISCORD_PUBLIC_KEY`.
...
  responses:
    ValidationErrorResponse:
      description: Input tidak valid (400).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: "customerName: Name is required; items: At least one item is required"
              details: { customerName: ["Name is required"] }
    UnauthorizedResponse:
      description: |
        Tidak terautentikasi / kredensial salah (401). Catatan: sebagian
        endpoint (confirm-transfer, payment-action, payments/expire, discord)
        membungkus `ValidationError` dengan status 401, sehingga `error.code`
        yang dikembalikan bisa `VALIDATION_ERROR` meski HTTP-nya 401.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            success: false
            error: { code: UNAUTHORIZED, message: "Authentication required" }
    NotFoundError:
      description: Resource tidak ditemukan (404).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          example:
            success: false
            error: { code: NOT_FOUND, message: "Product with id 'prod_abc' not found" }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta: { $ref: "#/components/schemas/Meta" }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, INTERNAL_ERROR]
            message: { type: string }
            details: {}
    Meta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

    # ── Enums ──
...
    Category:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        icon: { type: string, nullable: true }
        image: { type: string, nullable: true }
        sortOrder: { type: number }
        isActive: { type: boolean }
        parentId: { type: string, nullable: true }
...
    Setting:
      type: object
      properties:
        id: { type: string }
        key: { type: string }
        value: { type: string }
        type: { type: string }
        group_name: { type: string }
...
    ProductUpsertRequest:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: number }
        comparePrice: { type: number }
        unit: { type: string }
        stock: { type: integer }
        minStock: { type: integer }
        badge: { type: string }
        isActive: { type: boolean }
        isFeatured: { type: boolean }
        categoryId: { type: string }
        images:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
              alt: { type: string }
              sortOrder: { type: integer }
    CategoryUpsertRequest:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        icon: { type: string, nullable: true }
        image: { type: string, nullable: true }
        sortOrder: { type: integer }
        isActive: { type: boolean }
        parentId: { type: string, nullable: true }
...
    UpsertSettingRequest:
      type: object
      required: [key, value]
      properties:
        key: { type: string }
        value: { type: string }
        type: { type: string, default: text }
        group: { type: string, default: general }
...